% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin.specs.R
\name{bin.specs}
\alias{bin.specs}
\title{1D spectra equidistant binning}
\usage{
bin.specs(spec, ppm, bins = 0.005)
}
\arguments{
\item{spec}{NMR matrix, spectra in rows (matrix)}

\item{ppm}{Chemcial shift position of NMR matrix columns (vector)}

\item{bins}{Desired bin size (ppm)}
}
\description{
1D spectra equidistant binning
}
\author{
[Torben Kimhofer](https://tkimhofer.com)
}
