% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2d.R
\name{calibrate2d}
\alias{calibrate2d}
\title{Calibrate chemical shift vector for a NMR dimension}
\usage{
calibrate2d(
  spec2d,
  f2_ra = c(-0.1, 0.1),
  f1_ra = c(260, 3020),
  f2_fix = 0,
  f1_fix = 0
)
}
\arguments{
\item{spec2d}{list of 2D matrices with row and column names representing chemical shift values}

\item{f2_ra}{F2 range of expected peak of calibration signal}

\item{f1_ra}{F1 range of expected peak of calibration signal}

\item{f2_fix}{F1 axis position of calibrated signal}
}
\description{
Calibrate chemical shift vector for a NMR dimension
}
\section{}{
NA
}

\author{
torben.kimhofer@murdoch.edu.au
}
