% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convolve3cpp}
\alias{convolve3cpp}
\title{Perform matrix convolution}
\usage{
convolve3cpp(A, B)
}
\arguments{
\item{A}{Matrix A}

\item{B}{Matrix B}
}
\value{
A convolved with B
}
\description{
Perform matrix convolution
}
