% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d2cauchy_cpp}
\alias{d2cauchy_cpp}
\title{Generate 2D Cauchy distribution}
\usage{
d2cauchy_cpp(x, y, mu, gamma)
}
\arguments{
\item{x}{description of first dimensions (e.g. ppm values)}

\item{y}{description of second dimensions (e.g. Hz values)}

\item{mu}{distribution's center position using coordinate in x and y dimension}

\item{sigma}{distribution's spread (standard deviation) in x and y dimension}
}
\value{
Numeric vector of coordinates
}
\description{
Generate 2D Cauchy distribution
}
