% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2d.R
\name{.checkFiles}
\alias{.checkFiles}
\title{Select all 2d NMR spectral files with intact file directories}
\usage{
.checkFiles(datapath, procs_exp = 1, n_max = 10)
}
\arguments{
\item{datapath}{File path to J-res NMR experiment(s)}

\item{procs_exp}{Bruker processing experiment (usually 1)}

\item{n_max}{Maximum number of experiments to read}
}
\description{
Select all 2d NMR spectral files with intact file directories
}
\details{
Files needed: acqus, procs, proc2s, 2rr
}
\section{}{
NA
}

\author{
torben.kimhofer@murdoch.edu.au
}
