% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2d.R
\name{.chem_shift}
\alias{.chem_shift}
\title{Calculate chemical shift vector for a NMR dimension}
\usage{
.chem_shift(swidth, offset, si)
}
\arguments{
\item{swidth}{Sweep width / spectral width}

\item{offset}{Offset}

\item{si}{Number of data points of the respective NMR dimension}
}
\description{
Calculate chemical shift vector for a NMR dimension
}
\section{}{
NA
}

\author{
torben.kimhofer@murdoch.edu.au
}
