% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expandGrid_rcpp}
\alias{expandGrid_rcpp}
\title{List all combinations of two vectors}
\usage{
expandGrid_rcpp(x, y)
}
\arguments{
\item{x}{Vector A}

\item{y}{Vector B}
}
\value{
all combinations, iterating first on B, then on A
}
\description{
List all combinations of two vectors
}
