% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2d.R
\name{extract_pars}
\alias{extract_pars}
\title{Extract metadata from procs and proc2s files (JCAMP-JDX)}
\usage{
extract_pars(f_list)
}
\arguments{
\item{f_list}{List of directories and exp IDs (output checkFiles function)}
}
\description{
Extract metadata from procs and proc2s files (JCAMP-JDX)
}
\section{}{
NA
}

\author{
torben.kimhofer@murdoch.edu.au
}
