% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_multiplicity.R
\name{filter_multiplicity}
\alias{filter_multiplicity}
\title{Peak filter based on observed splitting pattern}
\usage{
filter_multiplicity(dfp, noise, dev.f1.hz = 0.2, dev.Jcoup.hz = 1.5)
}
\arguments{
\item{dfp}{Dataframe describing detected Jres features (see description)}

\item{noise}{Noise threshold}

\item{dev.f1.hz}{max diff in f1 (Hz), used for matching singlets to zero}

\item{dev.Jcoup.hz}{max diff Jcopuling symmetry}
}
\description{
dfp must have the following two columns: cent.f2 (center position f2 dimensions), cent.f1 (center position f1 dimension)
}
\author{
[Torben Kimhofer](https://tkimhofer.com)
}
