% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBBdim_rcpp}
\alias{getBBdim_rcpp}
\title{Find indices of points on axis that are closest after adding/subtracting a constant from a center point}
\usage{
getBBdim_rcpp(ppm, i, add)
}
\arguments{
\item{ppm}{Axis/ scale vector}

\item{i}{index for ppm, determining point where constand should be added/subtracted}

\item{add}{Constand to be added/subtracted (on ppm scale)}
}
\value{
indieces of ppm (lower, upper)
}
\description{
Find indices of points on axis that are closest after adding/subtracting a constant from a center point
}
