% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getBoundary}
\alias{getBoundary}
\title{Vector boundary finder when adding number to index}
\usage{
getBoundary(i, add_int, mat_end, type = "-")
}
\arguments{
\item{i}{index of vector}

\item{add_int}{index count added / subtracted from i}

\item{mat_end}{length of the vector}

\item{type}{addition or subtraction (specified by + or - symbol)}
}
\value{
all combinations, iterating first on B, then on A
}
\description{
Vector boundary finder when adding number to index
}
