% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lapOfG}
\alias{lapOfG}
\title{Calculate Laplacian of Gaussians to determine peak size / bounding box dimension}
\usage{
lapOfG(sub, f1hz, f2hz, cent_f1hz, cent_f2hz, sf, npix)
}
\arguments{
\item{sub}{Matrix with signal}

\item{f1hz}{Row scale of sub (Jres f1, expresed in Hz)}

\item{f2hz}{Col scale of sub (Jres f2, expressed in Hz)}

\item{cent_f1hz}{Location of peak maximum in f1 dimension}

\item{cent_f2hz}{Location of peak maximum in f2 dimension}

\item{sf}{Spectrometer frequency}

\item{npix}{SD parameter values used to create Gaussians - should span expected Jres signal fwhm range}
}
\value{
List, 1st element: Best matching Gaussian SD (peak size), 2nd element: Evaluation criterium for all tested SD
}
\description{
Calculate Laplacian of Gaussians to determine peak size / bounding box dimension
}
