% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matpow}
\alias{matpow}
\title{Power transform each matrix element}
\usage{
matpow(A, exp)
}
\arguments{
\item{exp}{power}

\item{Matrix}{A}
}
\value{
A^exp
}
\description{
Power transform each matrix element
}
