% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matsum}
\alias{matsum}
\title{Summation of two matrices of the same dimensions}
\usage{
matsum(A, B)
}
\arguments{
\item{A}{Matrix B}
}
\value{
A + B
}
\description{
Summation of two matrices of the same dimensions
}
