% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matsumup}
\alias{matsumup}
\title{Summation of all matrix elements to a double}
\usage{
matsumup(A)
}
\arguments{
\item{Matrix}{A}
}
\value{
sum(A)
}
\description{
Summation of all matrix elements to a double
}
