% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matsymminf2f1}
\alias{matsymminf2f1}
\title{Symmetrise signal using f2 as mirror axis and using min values}
\usage{
matsymminf2f1(A, idx_cent_row, idx_cent_col)
}
\arguments{
\item{A}{Matrix A}
}
\value{
summetrised A
}
\description{
Symmetrise signal using f2 as mirror axis and using min values
}
