% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minmaxMat}
\alias{minmaxMat}
\title{Min-max normalisation for matrix}
\usage{
minmaxMat(A)
}
\arguments{
\item{Matrix}{A}
}
\value{
min-max-scaled A (values range from 0-1)
}
\description{
Min-max normalisation for matrix
}
