% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise2D.R
\name{noise2D}
\alias{noise2D}
\title{Noise threshold estimation for 2D NMR spectrum}
\usage{
noise2D(jres, binsize = 0.02, sections = 32, sdlim = 3, probs = 0.1)
}
\arguments{
\item{jres}{J-coupling resolved NMR spectrum (df or matrix, t1 x t2)}

\item{binsize}{Bin size in f2 dimension (ppm)}

\item{sections}{window size f1, f2}

\item{sdlim}{Multiplier for SD to determine noise cutoff in each bin (usually value of three)}

\item{probs}{Quantile probability determining final noise cutoff for 2D}
}
\description{
Noise in JRES experiments is believed to be Rayleigh (RL) -like distributed (RL looks similar to a right-skewed normal dist). A good  noise cut-off is three times noise standard deviation (SD), however, due to right-skewness there will be few noise spikes  (Ref PhD thesis HM Parson, U Birmingham).
For each t1
1. bin spectum
2. divide binned spectrum in 32 equidistant sections
3. noise level is three times smallest SD found in all sections
4. take lower 10% value as noise cutoff for entire spectrum

Noise is likely to be different in different t1 areas, therefore a future application might focus on de-noising JRES/2D with location-specific noise values.
}
\author{
[Torben Kimhofer](https://tkimhofer.com)
}
