% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pickPeaks_rcpp}
\alias{pickPeaks_rcpp}
\title{Perform peak picking and get bounding box dimensions}
\usage{
pickPeaks_rcpp(jr, f1hz, f2ppm, noise, boundary, sf)
}
\arguments{
\item{jr}{Jres matrix, f1 in rows and f2 in cols}

\item{f1hz}{F1 scale of jr (Hz)}

\item{noise}{Intensity threshold for noise (no peaks are detected below this value)}

\item{boundary}{Initial bounding box estimate for determining peak size (one side estimate: x +/- boundary), this should be large enought to capture large signals (expressed in Hz)}

\item{sf}{Spectrometer frequency}

\item{f2hz}{F2 scale of jr (ppm)}
}
\value{
List of dataframes summarising the detected peaks/features
}
\description{
Perform peak picking and get bounding box dimensions
}
