% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotjres_overlay1D_bboxNull.R
\name{plotjres_overlay1D_bboxNull}
\alias{plotjres_overlay1D_bboxNull}
\title{Plotting of Jres including projections, inclusion of resp. 1D}
\usage{
plotjres_overlay1D_bboxNull(
  ds,
  t2.lim = c(3, 3.1),
  t1.lim = c(-10, 10),
  spec.1d,
  ppm.1d,
  z.probs = 0,
  SF = 600,
  bbox,
  title = "",
  pPeak = 0.6,
  tilt = F,
  addProjpp = NULL
)
}
\arguments{
\item{ds}{Jres with f1 in rows, f2 in cols}

\item{t2.lim}{Plotting range in f2 dimension}

\item{t1.lim}{Plotting range in f1 dimension}

\item{spec.1d}{1D NMR spectrum}

\item{ppm.1d}{Matching ppm scale for 1D NMR spec}

\item{z.probs}{Probability value for quantile function (to exlude lower intensities, e.g. noise)}

\item{SF}{Spec frequency}

\item{bbox}{dataframe of detected features for bounding box drawing (optional, can be set to NULL)}

\item{title}{Figure title}

\item{pPeak}{Probability threshold for drawing a feature's bounding box, part of bbox df and determined by SVM classification}

\item{addProjpp}{Vector of feature projections created by function *projFeat_f1_Tof2*}

\item{tile}{Should spec be tilted by 45 degree (depreciated)}
}
\description{
Plotting of Jres including projections, inclusion of resp. 1D
}
\author{
tkimhofer@gmail.com
}
