% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projFeat_f1Tof2.R
\name{projFeat_f1Tof2}
\alias{projFeat_f1Tof2}
\title{Create artificial 1D NMR by Jres feature projection}
\usage{
projFeat_f1Tof2(plist, jr, sf, lw.ppm)
}
\arguments{
\item{plist}{Dataframe of Jres peak summaries}

\item{jr}{Jres spectrum}

\item{sf}{Spectrometer frequency (MHz)}

\item{lwi}{estimated full widht at half max (fwhm) of 1D (usually TSP signal is fine, see spec.qc function)}
}
\description{
Function uses cauchy distribution (aka Lorentz distr) parameterised with max feature position on f2+f1 to determine location and 1D fwhm estimate to determine scale. Cauchy is scaled in intensity according to observed intensity of the resp. feature.
}
\author{
[Torben Kimhofer](https://tkimhofer.com)
}
