% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read2d.R
\name{read2d}
\alias{read2d}
\title{Read-in J-coupling resolved NMR spectra generated with a Bruker spectrometer}
\usage{
read2d(path, procs_exp = 1, n_max = 1000, interp = T)
}
\arguments{
\item{path}{File path to J-res NMR experiment(s)}

\item{procs_exp}{Bruker processing experiment (usually 1)}

\item{n_max}{Maximum number of experiments to read}

\item{inter}{bilinear interpolation}
}
\description{
Read-in J-coupling resolved NMR spectra generated with a Bruker spectrometer
}
\section{}{
NA
}

\author{
torben.kimhofer@murdoch.edu.au
}
