% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readBruker2d.R
\name{readBruker2d}
\alias{readBruker2d}
\title{Read-in J-coupling resolved NMR spectra generated with a Bruker spectrometer}
\usage{
readBruker2d(datapath, n)
}
\arguments{
\item{datapath}{File path to J-res NMR experiment(s)}

\item{n}{Maximum number of experiments to read}
}
\description{
Read-in J-coupling resolved NMR spectra generated with a Bruker spectrometer
}
\author{
tkimhofer@gmail.com
}
