% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_multiplicity.R
\name{filter_multiplicity}
\alias{filter_multiplicity}
\title{Peak filter based on observed splitting pattern}
\usage{
filter_multiplicity(dfp, noise, f1.var = 0.5)
}
\arguments{
\item{dfp}{peak list}

\item{noise}{noise level}

\item{f1.var}{max diff in f1 (Hz), used for matching singlets to zero}
}
\description{
Peak filter based on observed splitting pattern
}
\author{
[Torben Kimhofer](https://tkimhofer.com)
}
